fix issues in repo so that prepack can run on the machine and then the result be ran in browser
