+++ js/ast/abstraction.js
@@ -1,0 +1,13 @@
+define('ast/abstraction', function() {
+    class Abstraction {
+      /**
+       * param here is the name of the variable of the abstraction. Body is the
+       * subtree  representing the body of the abstraction.
+       */
+      constructor(param, body) {
+        this.param = param;
+        this.body = body;
+      }
+    }
+    return Abstraction;
+});
