+++ js/ast/assign.js
@@ -1,0 +1,9 @@
+define('ast/assign', function() {
+    class Assign {
+      constructor(param, body) {
+        this.param = param;
+        this.body = body;
+      }
+    }
+    return Assign;
+});
