+++ js/ast/change.js
@@ -1,0 +1,9 @@
+define('ast/change', function() {
+    class Change {
+      constructor(param, body) {
+        this.param = param;
+        this.body = body;
+      }
+    }
+    return Change;
+});
