+++ js/ast/constant.js
@@ -1,0 +1,8 @@
+define('ast/constant', function() {
+    class Constant {
+      constructor(value) {
+        this.value = value;
+      }
+    }
+    return Constant;
+});
