+++ js/ast/deprecation.js
@@ -1,0 +1,8 @@
+define('ast/deprecation', function() {
+    class Deprecation {
+      constructor(term) {
+        this.term = term;
+      }
+    }
+    return Deprecation;
+});
