+++ js/ast/deref.js
@@ -1,0 +1,8 @@
+define('ast/deref', function() {
+    class Dereference {
+      constructor(term) {
+        this.term = term;
+      }
+    }
+    return Dereference;
+});
