+++ js/ast/operation.js
@@ -1,0 +1,9 @@
+define('ast/operation', function() {
+    class Operation {
+      constructor(type, name) {
+        this.type = type;
+        this.name = name;
+      }
+    }
+    return Operation;
+});
