+++ js/ast/propagation.js
@@ -1,0 +1,7 @@
+define('ast/propagation', function() {
+    class Propagation {
+      constructor() {
+      }
+    }
+    return Propagation;
+});
