+++ js/ast/provisional-constant.js
@@ -1,0 +1,8 @@
+define('ast/provisional-constant', function() {
+    class ProvisionalConstant {
+      constructor(term) {
+        this.term = term;
+      }
+    }
+    return ProvisionalConstant;
+});
