+++ js/ast/recursion.js
@@ -1,0 +1,9 @@
+define('ast/recursion', function() {
+    class Recursion {
+      constructor(param, body) {
+        this.param = param;
+        this.body = body;
+      }
+    }
+    return Recursion;
+});
