+++ js/gc.js
@@ -1,11 +1,5 @@
-define(function(require) {
-
-	var Group = require('group');
-	var Contract = require('nodes/contract');
-	var Promo = require('nodes/promo');
-	var Recur = require('nodes/recur');
-	var Abs = require('nodes/abs');
-	var Weak = require('nodes/weak');
+define(['group', 'nodes/contract', 'nodes/promo', 'nodes/recur', 'nodes/abs', 'nodes/weak'],
+	function(Group, Contract, Promo, Recur, Abs, Weak) {
 
 	class GC {
