+++ js/nodes/deref.js
@@ -1,6 +1,4 @@
-define(function(require) {
-
-	var Node = require('node');
+define(['node'], function(Node) {
 
 	class Deref extends Node {
