+++ js/nodes/prov.js
@@ -1,11 +1,8 @@
-define(function(require) {
+define(['node', 'token', 'nodes/mod', 'nodes/const', 'link'],
+	function(Node, Token, Mod, Const, Link) {
 
-	var Node = require('node');
-	var CompData = require('token').CompData();
-	var RewriteFlag = require('token').RewriteFlag();
-	var Mod = require('nodes/mod');
-	var Const = require('nodes/const');
-	var Link = require('link');
+	var CompData = Token.CompData();
+	var RewriteFlag = Token.RewriteFlag();
 
 	class Prov extends Node {
