+++ js/nodes/unop.js
@@ -1,13 +1,10 @@
-define(function(require) {
+define(['node', 'token', 'link', 'box-wrapper', 'nodes/promo', 'nodes/const', 'op'],
+	function(Node, Token, Link, BoxWrapper, Promo, Const, Op) {
 
-	var Node = require('node');
-	var CompData = require('token').CompData();
-	var RewriteFlag = require('token').RewriteFlag();
+	var CompData = Token.CompData();
+	var RewriteFlag = Token.RewriteFlag();
 	var Link = require('link');
-	var BoxWrapper = require('box-wrapper');
-	var Promo = require('nodes/promo');
-	var Const = require('nodes/const');
-	var UnOpType = require('op').UnOpType;
+	var UnOpType = Op.UnOpType;
 
 	class UnOp extends Node {
