+++ main.js
@@ -7,16 +7,6 @@ requirejs.config({
 	//config is relative to the baseUrl, and
 	//never includes a ".js" extension since
 	//the paths config could be for a directory.
-	bundles: {
-        'parser/ast': ['ast/abstraction', 'ast/application', 'ast/identifier'
-        			 , 'ast/constant', 'ast/operation', 'ast/unary-op'
-        			 , 'ast/binary-op', 'ast/if-then-else', 'ast/recursion'
-        			 , 'ast/provisional-constant', 'ast/change', 'ast/assign', 'ast/propagation'
-        			 , 'ast/deprecation', 'ast/deref'],
-        'group': ['group', 'box'],
-        'term': ['term', 'box-wrapper'],
-        'nodes/der': ['nodes/der', 'nodes/var'],
-   	},
 	paths: {
 		d3: '../bower_components/d3/d3',
 		"dot-checker": '../bower_components/graphviz-d3-renderer/dist/dot-checker',
@@ -24,7 +14,8 @@ requirejs.config({
 		worker: '../bower_components/requirejs-web-workers/src/worker',
 		renderer: '../bower_components/graphviz-d3-renderer/dist/renderer',
 		jquery: '../bower_components/jquery/dist/jquery',
-		"goi-machine": './goi-machine',
+		// "goi-machine": './goi-machine',
+		"goi-machine": '../../tas458/lib/goiMachine/v2/main.requirejs',
 	},
 	shim: {
 		jquery: {
